Project renamed and blazor projects references updated
