+++ RichTextEditor.sln
@@ -1,10 +1,8 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29519.87
+# Visual Studio Version 17
+VisualStudioVersion = 17.4.33205.214
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Blazored.TextEditor", "src\Blazored.TextEditor\Blazored.TextEditor.csproj", "{A76F9184-D5C2-42AE-9BFA-BDF33A3B1434}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BlazorServerSide", "samples\BlazorServerSide\BlazorServerSide.csproj", "{AEA7320C-E669-4009-B877-20E49D1D1687}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BlazorClientSide", "samples\BlazorClientSide\BlazorClientSide.csproj", "{59EDB617-CA0F-472B-B107-A4E1694A3691}"
@@ -19,16 +17,14 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		README.md = README.md
 	EndProjectSection
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RichTextEditor", "src\RichTextEditor\RichTextEditor.csproj", "{CE5AC536-F87C-4082-925E-660CD0525244}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{A76F9184-D5C2-42AE-9BFA-BDF33A3B1434}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A76F9184-D5C2-42AE-9BFA-BDF33A3B1434}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A76F9184-D5C2-42AE-9BFA-BDF33A3B1434}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A76F9184-D5C2-42AE-9BFA-BDF33A3B1434}.Release|Any CPU.Build.0 = Release|Any CPU
 		{AEA7320C-E669-4009-B877-20E49D1D1687}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{AEA7320C-E669-4009-B877-20E49D1D1687}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{AEA7320C-E669-4009-B877-20E49D1D1687}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -37,6 +33,10 @@ Global
 		{59EDB617-CA0F-472B-B107-A4E1694A3691}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{59EDB617-CA0F-472B-B107-A4E1694A3691}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{59EDB617-CA0F-472B-B107-A4E1694A3691}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CE5AC536-F87C-4082-925E-660CD0525244}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CE5AC536-F87C-4082-925E-660CD0525244}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CE5AC536-F87C-4082-925E-660CD0525244}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CE5AC536-F87C-4082-925E-660CD0525244}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
