+++ samples/BlazorClientSide/BlazorClientSide.csproj
@@ -13,7 +13,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\src\Blazored.TextEditor\Blazored.TextEditor.csproj" />
+    <ProjectReference Include="..\..\src\RichTextEditor\RichTextEditor.csproj" />
   </ItemGroup>
 
 </Project>
