+++ samples/BlazorClientSide/Pages/Index.razor
@@ -1,5 +1,5 @@
 ﻿@page "/"
-@using Blazored.TextEditor
+@using RichTextEditor
 
 <BlazoredTextEditor @ref="@QuillHtml">
     <ToolbarContent>
