+++ samples/BlazorServerSide/BlazorServerSide.csproj
@@ -5,7 +5,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\src\Blazored.TextEditor\Blazored.TextEditor.csproj" />
+    <ProjectReference Include="..\..\src\RichTextEditor\RichTextEditor.csproj" />
   </ItemGroup>
 
 </Project>
