+++ samples/BlazorServerSide/Pages/_Host.cshtml
@@ -26,8 +26,8 @@
 
     <!-- Quill library -->
     <script src="https://cdn.quilljs.com/1.3.6/quill.js"></script>
-    <script src="_content/Blazored.TextEditor/quill-blot-formatter.min.js"></script>
-    <script src="_content/Blazored.TextEditor/Blazored-BlazorQuill.js"></script>
+    <script src="_content/RichTextEditor/quill-blot-formatter.min.js"></script>
+    <script src="_content/RichTextEditor/Blazored-BlazorQuill.js"></script>
 
 </body>
 </html>
