+++ src/RichTextEditor/RichTextEditor.csproj
@@ -3,20 +3,21 @@
   <PropertyGroup>
     <TargetFramework>netstandard2.0</TargetFramework>
     <RazorLangVersion>3.0</RazorLangVersion>
-    <RootNamespace>Blazored.TextEditor</RootNamespace>
+    <RootNamespace>RichTextEditor</RootNamespace>
 
-    <PackageId>Blazored.TextEditor</PackageId>
-    <Version>1.0.8</Version>
-    <Authors>Michael Washington</Authors>
+    <PackageId>RichTextEditor</PackageId>
+    <Version>1.0.0</Version>
+    <Authors>Sammy Guadarrama Chávez, Michael Washington</Authors>
     <Description>Rich text editor for Blazor applications - Uses Quill JS</Description>
     <Copyright>Copyright 2022 (c) Michael Washington. All rights reserved.</Copyright>
+    <Copyright>Copyright 2023 (c) Sammy Guadarrama Chávez. All rights reserved.</Copyright>
     <PackageLicenseExpression>MIT</PackageLicenseExpression>
-    <PackageProjectUrl>https://github.com/Blazored/TextEditor</PackageProjectUrl>
-    <PackageIconUrl>https://avatars1.githubusercontent.com/u/39194025</PackageIconUrl>
-    <RepositoryUrl>https://github.com/Blazored/TextEditor</RepositoryUrl>
+    <PackageProjectUrl>https://github.com/SammyGCh/RichTextEditor</PackageProjectUrl>
+    <!--<PackageIconUrl>https://avatars1.githubusercontent.com/u/39194025</PackageIconUrl>-->
+    <RepositoryUrl>https://github.com/SammyGCh/RichTextEditor</RepositoryUrl>
     <PackageTags>Blazor Rich Text Editor Quill QuillJS Blazored Razor Components</PackageTags>
     <Company />
-    <PackageReleaseNotes>Adds Id to Div</PackageReleaseNotes>
+    <PackageReleaseNotes>Allows to get the editor text content in OnChange event.</PackageReleaseNotes>
   </PropertyGroup>
 
   <ItemGroup>
