+++ src/RichTextEditor/icon.png
